Make HoodieFlinkCompactor could read parameter by specifying options which like --props and --hoodie-conf.
